pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = "Convert PNG images to SVG hexagon wallpapers"
authors = ["jeangnamien <jean.gnamien@master.centralelille.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^12.0.0"
marimo = "^0.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
black = "^25.11.0"
pylint = "^4.0.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"